requirements.txt
Django==3.1.2
xlrd==1.2.0
Pillow
djangorestframework
django-crontab==0.7.1
django-uuslug==1.2.0
django-allauth==0.42.0
requests==2.24.0
django-cleanup==5.1.0
python-decouple==3.3
uvicorn==0.19.0
redis==3.5.3
django-ajax-selects==2.0.0
django-background-tasks==1.2.5
celery==5.0.1
six==1.15.0
django-admin-rangefilter==0.11.2
django-simple-history==3.0.0